refactor(upstream): extract pool/validate.rs (lab-kvji.12.4 step 5)
Move validate_upstream_config + its 8 validate_* tests into pool/validate.rs;
move estimate_response_size into pool/helpers.rs. Behavior unchanged. 1423 green.

Bead: lab-kvji.12.4
